+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+### 0.2.0
+* fix config.rb requiring
+* broadcast async connection iteration
+
 ### 0.1.0
 * initial version